Refactor/불필요한 Styled컴포넌트 제거 및 interface 상속
