+++ client/src/common/style/Containers.styled.tsx
@@ -17,13 +17,14 @@ export const ContentContainer = styled.div<{size:number|string, direction:string
 `;
 
 export const ColumnItemWrapper = styled.div<{size:number|string, gap:number}>`
-    width: ${(props)=> typeof props.size === 'string' ? props.size : `${props.size}px`};
-    ${FlexColumnCenter}
+    width: ${(props) => typeof props.size === 'string' ? props.size : `${props.size}px`};
+    display: flex;
+    flex-direction: column;
     gap: ${(props)=> props.gap}px;
     background-color: transparent;
 `;
 
-export const RowItemWrapper = styled.div`
-  ${FlexCenter};
-  gap: 5px;
+export const RowItemWrapper = styled.div<{gap:number}>`
+  display: flex;
+  gap: ${(props)=> props.gap}px;
 `;
