+++ client/src/common/style/Images.styled.tsx
@@ -1,0 +1,8 @@
+import { css, styled } from 'styled-components';
+
+export const Favicon = styled.img`
+    width: 16px;
+    height: 16px;
+    margin: 2px 6px 0 0;
+    border: solid 2px red;
+`
